Refactor generate stack to live in transforms.py
